+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.23.2 (unreleased)
 
 - Speed up combined_headers in ssdp_listener (@bdraco, #115)
+- Add handling of broken SSDP-headers (#116)
 
 
 0.23.1 (2021-12-18)